finish moving dp related code
